♻️ Separate code into multiple files
